--- backend/.env.example
@@ -1,10 +1,0 @@
-# Database Configuration
-# You can use individual variables (Recommended for Docker)
-DB_USER=root
-DB_PASSWORD=secret
-DB_HOST=localhost
-DB_PORT=3306
-DB_NAME=stacknote
-
-# OR you can use a full connection string
-# MYSQL_DATABASE_URL="mysql+pymysql://user:password@localhost:3306/dbname"
