+++ backend/app/routers/work_space.py
@@ -2,7 +2,7 @@ from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.orm import Session
 
 from app.core.database import get_db
-from app.models import WorkSpace
+from app.workspace.model import WorkSpace
 from app.schemas.work_space import (
     WorkspaceRequest,
     WorkspaceResponse,
