+++ docker-compose.yml
@@ -3,21 +3,21 @@
 services:
   backend:
     build: ./backend
-    container_name: stacknote-b_feat-4-backend-1
+    container_name: stacknote-backend
     ports:
       - "8000:8000"
     volumes:
       - ./backend/sql_app.db:/app/sql_app.db
       - ./backend/app:/app/app        # 코드 수정 시 컨테이너에 즉시 반영되도록 추가
     env_file:
-      - ./backend/.env
+      - .env
     depends_on:
       - db
     restart: always
 
   voice-backend:
     build: ./voice-backend
-    container_name: stacknote-b_feat-4-voice-backend-1
+    container_name: stacknote-voice-backend
     ports:
       - "8001:8000"
     volumes:
@@ -26,7 +26,7 @@ services:
 
   frontend:
     build: ./frontend
-    container_name: stacknote-b_feat-4-frontend-1
+    container_name: stacknote-frontend
     ports:
       - "5173:5173"
     volumes:
@@ -38,12 +38,12 @@ services:
 
   db:
     image: mysql:8.4
-    container_name: stacknote-b_feat-4-db-1
+    container_name: stacknote-db
     environment:
       MYSQL_ROOT_PASSWORD: ${DB_PASSWORD}
       MYSQL_DATABASE: ${DB_NAME}
     ports:
-      - "3306:3306"
+      - "3307:3306"
     volumes:
       - db_data:/var/lib/mysql
     restart: always
